# public/docs/api-docs.yaml
openapi: 3.0.0
info:
  description: "[TrungPhuNA] Demo document API"
  version: "1.0.0"
  title: "Core Dự án"
  termsOfService: "http://123code.net/"
  contact:
    email: "codethue94@gmail.com"

components:
  securitySchemes:
    ApiNetworkAuth:
      type: apiKey
      in: header
      name: X-Network-Token
    ApiPortAuth:
      type: apiKey
      in: header
      name: X-Port-Type

security:
  - ApiNetworkAuth: []
  - ApiPortAuth: []

host: "code-duan.abc:8888"
basePath: "/api"

servers:
  - url: http://code-duan.abc:8888/api
  - url: https://cms.123code.net/api

# Tags đại diện cho các table
tags:
  - name: "products"
    description: "Sản phẩm"

#schemes: [http, https]

paths:
  /products:
    get:
      security:
        - ApiNetworkAuth: []
        - ApiPortAuth: []
      tags:
        - "products"
      summary: "Danh sách sản phẩm"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - name: "limit"
          in: "path"
          description: "Số bản ghi cần lấy trên 1 page"
          required: true
          type: "integer"
          format: "int64"
        - name: "page"
          in: "path"
          description: "Số trang"
          required: false
          type: "integer"
      responses:
        "404":
          description: "Không tồn tại đường dẫn"

  /products/{slug}:
    get:
      tags:
        - "products"
      summary: "Lấy thông tin sản phẩm theo [Slug]"
      description: "Returns a product"
      operationId: "getProductBySlug"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - name: "slug"
          in: "path"
          description: "Slug sản phẩm"
          required: true
          type: "string"
      responses:
        "200":
          description: "Success"
        "404":
          description: "Đường dẫn API không tồn tại"

definitions:
  Category:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Category"